requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.0
spacy==3.7.2
python-dotenv==1.0.0
pandas==2.1.3
email-validator==2.1.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
websockets==12.0 